deleted cracker.rb and crack_test.rb because they werent doing anything, also modified FileWriter to warn use if file name given already exists before overwriting

+++ lib/enigma/file_writer.rb
@@ -7,9 +7,24 @@ class FileWriter
 	end
 
 	def write_encryption(encryption)
-		file = File.open("#{file_name}", 'w')
-		file.write(encryption)
-		file.close
+		if File.exist?(file_name)
+			puts "WARNING: A file named #{file_name} already exists. Would you like to overwrite it? (y/n)"
+			overwrite = $stdin.gets.chomp
+
+			if overwrite == "y"
+				file = File.open("#{file_name}", 'w')
+				file.write(encryption)
+				file.close
+			else
+				puts "Try again with a different file name. Bye!"
+				exit
+			end
+
+		else
+			file = File.open("#{file_name}", 'w')
+			file.write(encryption)
+			file.close
+		end
 	end
 
 end
